+++ start_rfid.sh
@@ -3,4 +3,4 @@
 export LD_LIBRARY_PATH=$LD_LIBRARY_PATH:/opt/feig
 
 cd plugins/rfid/device
-java -jar rfid.jar
+java -jar rfid.jar logtoconsole=true loglevel=debug